Build of persisted state to run in parallel with work on dirty state.
Enable a build of the persisted state to run in parallel to work
performed on the dirty state by splitting the index, the resource set,
the content provider, and the request manager into persisted and dirty
components.

Introduces a limited scope dirty build so that validations are performed
for dirty editors and the effect of changes in one dirty editor can be
seen in other dirty editors.

+++ org.eclipse.xtext/src/org/eclipse/xtext/build/IncrementalBuilder.java
@@ -51,7 +51,6 @@ import org.eclipse.xtext.resource.persistence.StorageAwareResource;
 import org.eclipse.xtext.service.OperationCanceledManager;
 import org.eclipse.xtext.util.CancelIndicator;
 import org.eclipse.xtext.util.RuntimeIOException;
-import org.eclipse.xtext.validation.CheckMode;
 import org.eclipse.xtext.validation.IResourceValidator;
 import org.eclipse.xtext.validation.Issue;
 import org.eclipse.xtext.workspace.IProjectConfig;
@@ -350,7 +349,7 @@ public class IncrementalBuilder {
 			if (resourceValidator == null) {
 				return true;
 			}
-			List<Issue> validationResult = resourceValidator.validate(resource, CheckMode.ALL,
+			List<Issue> validationResult = resourceValidator.validate(resource, context.getValidationMode(),
 					request.getCancelIndicator());
 			return request.getAfterValidate().afterValidate(resource.getURI(), validationResult);
 		}
@@ -359,6 +358,9 @@ public class IncrementalBuilder {
 		 * Generate code for the given resource
 		 */
 		protected void generate(Resource resource, BuildRequest request, Source2GeneratedMapping newMappings) {
+			if (context.isDirtyBuild()) {
+				return;
+			}
 			IResourceServiceProvider serviceProvider = getResourceServiceProvider(resource);
 			Set<URI> previous = newMappings.deleteSource(resource.getURI());
 			URIBasedFileSystemAccess fileSystemAccess = createFileSystemAccess(serviceProvider, resource);
